Add code for dynamic word list
Also overhauls the file structure for storing audio. Having it all in the src folder meant it was going to be next to impossible to dynamically import all the audio, and moving it to the public folder meant I could basically just link to it as you'd expect. The code for the dynamic audio getter/area, was written with assistance from ChatGPT, of course I needed to modify it for my project though.