Complete Answer highlighting
Now allows the user to select multiple options, and then adds the correct class based on what the word is(correct/incorrect).
This is done by adding the index of the mapped button to an object, recording whether or not the user has clicked on it. If they have, and it's incorrect it gets the `flashcard-incorrect` class added on to the end (CSS styling it red for now) if it's correct, it adds the `flashcard-correct` class instead.
Doing it this way, (as an object, with keys and values) allows me to only apply the class to that one button, instead of ALL buttons as it was previously.