Add audio tag to flashcard word
In response to feedback, I've added the audio tag underneath the current word in the flashcard game.
Makes use of another `useEffect` to use `useRef` to update the value of the audio tag and actually have it playable. Trying this with useState would update the 'src' of the audio to the correct path, but would not actually update the audio in a way that was playable.