Make use of localStorage to set user's prefs
Should now set localStorage settings to store the user's preferred view.
By default, it'll be set to the list view, but if the user changes it, it should update localstorage and then read that value the next time the user loads the page, whether that be refreshing the page, navigating between areas in the app, or a new browser session (without clearing local storage).