Fix tooltips not displaying certain characters
Added a check to ensure that the tooltip for special characters/combinations such as  əl, əm, or ən get added.
This is done by checking to see if the sound exists, OR if it exists if you replace the special character with an empty string. This doesn't change the output of those characters, it just makes it a more flexible match.