Fix CSS tweaks on basics page
Items in the sound list are now aligned. Additionally, padding was added to the bottom of each sound entry. The (backsign) in the heading of Basic Concepts is now separated from the actual backsign character by a span tag, allowing me to style just the backsign instead of including the (backsign) in it